requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
decorator==5.1.1
Django==4.1.6
djangorestframework==3.14.0
executing==1.2.0
ipdb==0.13.11
ipython==8.10.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
PyJWT==2.6.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
wcwidth==0.2.6
